requirements.txt
xarm-python-sdk
pyyaml
requests
pyserial
websocket-client  # for future Moonraker WebSocket upgrade